Major restructuring of the port monitor server-side component.
The initial design relied on the Xcv port functionality to set job specific
parameters for a particular port, however, it turned out that
communicating via Xcv ports required a special privilage that only
administrators have by default. This presented a problem since job specific
parameters are meant to be set by applications running under a limited user
account.

To overcome the problem I decided to use named pipes to set job specific
parameters, which should in theory allow everything to function under a limited
user account. While it would've been possible to implement all this in the
server dll the prospect of debugging a pipe server with the Windows spooler
sitting in the middle (and getting in the way) didn't appeal to me.

Instead I decided to move pretty much all the existing functionality from the
server dll to a new service, with the server dll essentially acting as a proxy
for the service. This redesign also makes it possible for applications to
directly communicate the service via named pipes.

Moving most of the functionality into the service also makes it much easier to
debug and allows me make use of the .NET framework.